+++ src/main/java/com/hbm/inventory/recipes/loader/SerializableRecipe.java
@@ -82,6 +82,7 @@ public abstract class SerializableRecipe {
         recipeHandlers.add(ChemicalPlantRecipes.INSTANCE);
         recipeHandlers.add(PUREXRecipes.INSTANCE);
         recipeHandlers.add(FusionRecipes.INSTANCE);
+        recipeHandlers.add(PlasmaForgeRecipes.INSTANCE);
         recipeHandlers.add(PrecAssRecipes.INSTANCE);
     }